Added some comments, removed unused code, formatted some code. moved some code into  tcproxy-core
