+++ tcproxy-cli/src/commands/data_packet.rs
@@ -2,28 +2,34 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use bytes::BytesMut;
-use tracing::debug;
 use tcproxy_core::{Command, Result};
+use tracing::debug;
 use uuid::Uuid;
 
 use crate::ClientState;
 
+/// issued when server receives new data packet.
 pub struct DataPacketCommand {
-  connection_id: Uuid,
-  buffer: BytesMut,
-  buffer_size: u32,
-  state: Arc<ClientState>,
+    connection_id: Uuid,
+    buffer: BytesMut,
+    buffer_size: u32,
+    state: Arc<ClientState>,
 }
 
 impl DataPacketCommand {
-  pub fn new(connection_id: Uuid, buffer: BytesMut, buffer_size: u32, state: &Arc<ClientState>) -> Self {
-    Self {
-      connection_id,
-      buffer,
-      buffer_size,
-      state: state.clone(),
+    pub fn new(
+        connection_id: Uuid,
+        buffer: BytesMut,
+        buffer_size: u32,
+        state: &Arc<ClientState>,
+    ) -> Self {
+        Self {
+            connection_id,
+            buffer,
+            buffer_size,
+            state: state.clone(),
+        }
     }
-  }
 }
 
 #[async_trait]
