+++ tcproxy-cli/src/commands/remote_disconnected.rs
@@ -8,6 +8,7 @@ use uuid::Uuid;
 
 use crate::ClientState;
 
+/// issued when remote socket disconnects from server.
 pub struct RemoteDisconnectedCommand {
     connection_id: Uuid,
     state: Arc<ClientState>,
