+++ tcproxy-cli/src/lib.rs
@@ -5,7 +5,11 @@ mod ping_sender;
 mod frame_reader;
 mod frame_writer;
 mod args;
+mod app;
+mod console_updater;
 
+pub use app::*;
+pub use console_updater::*;
 pub use args::*;
 pub use frame_writer::*;
 pub use frame_reader::*;
